requirements.txt
praw==7.7.0
openai==0.27.0
python-wordpress-xmlrpc==2.3
python-dotenv==0.19.2
